reorganize tests into an icky macro invocation
